+++ src/geometry_utils.cpp
@@ -153,7 +153,7 @@ bool plane_intersect(
   if (!intersect_backfacing && glm::dot(normal,r.getDirection()) >= 0) 
     return 0; // hit the backside
 
-  float t = numer / denom;
+  double t = numer / denom;
 
   if (t > EPSILON && t < h.getT()) {
     h.set(t,normal);
